requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-reversion==3.0.9
django-reversion-compare==0.14.0
djangorestframework-jwt==1.11.0
django-rest-swagger==2.2.0
beautifulsoup4==4.7.0
requests==2.21.0
django-cors-headers==3.2.0
mysqlclient==2.0.3